Added a toolsmithing executable
This happens to me all the time: I get an idea for improving my tools,
but I cannot execute it immediatelly. Either I am in a Pomodoro or it is
too big and I have to do some research. Just getting an idea and not
doing anything is stressful -- your brain tries to hold on to it, while
you are trying to focus on something else. It is a lot better if you
just write it down and continue, getting back to it eventually.

Thus, this is my TODO, my wishlist, all the things I would like to do
eventually, version controlled.