git: browse all commits from yesterday
We have this trying in our company, where we have to read all commits
everybody made every day. I want to read them in the terminal and
I want to have links to GitHub, if I want to comment on something.

This is a very hacky, ad-hoc solution to this problem. I've tested it
only on GNU and I'm certain that lesskey does not work on the Mac. If
you ever wonder, you run

$ git catchup

This will open less, which in turn will show an index of all commits made
in the previous 24 hours. You can navigate to the previous and next
commits with h and l (vim-like bindings, that's where lesskey kicks in).
You can go back to the index with 0.

UX is far from perfect and probably there is something else that has this
functionality baked in, but it was a nice learning experience.

And just to highlight, the point was the link to GitHub.