vim: add a paging function
You know how you have it when you type :map and can't search into it?
Enter the Page command. Just do

  :Page map

The result will be open in a new buffer where you can do whatever you
want with it.

Now for a funny Yak shaving story:

I was looking at the TODO file in this project and preparing to delete
it. There was an entry to define a relativenumber mapping. I knew that
there was one already in my setup, but I had no idea where it came from.
I tried doing :verbose map, but got frustrated at the lack of search. I
found an answer in stackoverflow how to capture the output, but I wanted
to abstract that. 30 minutes later this came into life.

That stuff tends to happen quite often to me when dotfiles are involved.