tools: hivemind errors in non-opened files
Visual Studio Code's <F8> is all and great, but it doesn't jump to
errors in files that have not been opened, because Visual Studio Code
does not see them.

Not. Any. More.

This script helps make VSCode aware of those errors so the process could
be sped up quite significantly.

Note that VSCode will eventually support this, but until then - perl to
the rescue.