tmux: improve mouse selection
If you scroll back and select something, the moment you release the
mouse key, tmux scrolls down to the bottom because it exits copy mode.

Not any more. It instead remains where it is.

Additionally, if you're at the bottom of the scroll buffer, it exits
automatically, as you probably weren't in copy anyway.