vim: disable hlsearch
OK, this is big. I feel I am a bit slow in my buffer navigation, because
I hate using /, because it leaves a highlight mark and then I need to do
a :noh otherwise I get distracted. By first disabling hlsearch and then
sacrificing the incredibly useful <C-h> to flip it, I'd like to see if
this will improve my navigation. I have a couple of other ideas, but
they can wait for when I have the time.