tools: redo track-hivemind in Ruby
Alright, time for a long one.

scripts/track-hivemind got broken when we moved from webpacker to
vanilla webpack because the output was different. I wanted to fix it,
but it was annoying to do so, as it was hard to test the changes I'm
making. I needed to write a unit test, but it was a weird Perl script,
and now I faced a conundrum.

I made a decision a while ago to do more dotfile scripting in Perl. This
was purely for hedonistic reasons – I just enjoy writing Perl. But I run
into two problems:

* Some "advanced" techniques, like coroutines and threads, are awkward
  in Perl. They are doable, but I don't know how to do them and will
  need to learn.

* Non-trivial project setup (e.g. having your tests and your code in the
  same shebang runnable file) is tricky in Perl. It's doable, but I
  don't know how to do it, and I will need to learn.

Either I have to spend a lot of time learning auxiliary Perl skills,
which is fine, but time-consuming in a I'd-love-to-do-it-but-I-do-not-
have-the-time way. So I had to make a choice, and the choice is to keep
doing the same thing, but in Ruby, which is the modern Perl anyway. I
loose the learning fun, but as a bonus:

- I'm so much nimbler when doing this in Ruby
- I get to start a new Ruby codebase I can use to testbed Ruby tools I'm
  potentially gonna tools (or stuff that's already written like sorbet
  and solargraph)
- Ruby is fun too! I don't get to do it enough at work for some weird
  reason.

This, this is now moving to Ruby. I'm gonna start a Swamp-like thing,
but in Ruby. I chose to call it "mire", because that's a synonym and
also a fun word. Yeah, I'm lazy that way.

Like or retweet if you enjoyed reading.