nvim: add a basic command palette
I find it hard to remember some of my customizations, so I'd like to
have them searchable. I'd also like to separate the internal ones from
the ones I intend to use. Notably, the rubocop stuff like
RubocopDisableLine and RubocopDisableInProject are a bit hard to
remember and _very_ annoying to type. This needs a solution.

Enter the palette. It's semi-inspired from the command palette in visual
studio code (which, of course is based on stuff in its predecessors). It
uses:

- Telescope to visualise the available commands
- Mire in order to grep the vim files for documented commands
- Some neovim goodies in order to make stuff smooth.