mire: a super rudimentary continuous test runner
There's a very long way to go with this, but I wanted to have the basic
pieces and commit them to git. It doesn't do much apart from listening
to changes and running the corresponding spec with the RSpec beholder
attached. It can be significantly smarted, but I wanted to get this out
of the way before I zoom into other things (like figuring out how to
indicate failures in Vim or talking to it).

On the plus side, there is a bug(?) in the RSpec beholder that took me
two hours to figure because I was looking at the wrong place (towards
all the fiber-related code).

Oh, and I'm calling it wire, because I remember reading a book on
continuous testing about 10 years ago that had a SMART-like acronym
that was "FIRE", the phrase "Mire Fire" is absurd, and I'm quickly
descending into buffoonery.