mire: extend fire with coverage and wire to vim
This is quite exciting!

First, Fire now collects coverage data. It can be quite a chunky piece
of data, so it does it in a very rube-goldberg kind of way. Bottom line,
though, is that it can get it.

Second, Vim can now listen for fire events and do stuff when they
happened. In particular, it displays both the errors and the coverage
with extmarks.

There's quite a way to go:

- All coverage data goes to vim, while it should get only what it needs.

- It absolutely looks like a christmas tree right now. Saner
  highlighting is necessary.

- Simplecov still dumps some files in /tmp which can be monkey-patched
  away.