vim: add a test harness
- Add Vader to test vimscript
- Add an example test with vader
- Add a fat helper to invoke s: functions from tests
- Add an example test with plenary
- Get plugin/runner.vim to be able to run them
- Add a Makefile for good measure

Neither are very useful tests, but I've wanted to have the capability to
write tests, and I want to have it ready next time I need it.

Interestingly enough, I tend to find it not that useful, because in the
context of this, I need tests mostly for quick feedback, and I usually
manage to set up other means to have quick feedback in Vim.