nvim: fix a weird one
For some reason vimrc gets loaded after the filetype plugins. I have
yet to figure out why (plug.vim?), but when I open a lua file directly
from command line (as opposed to starting a vim and then finding it), I
get an error because it cannot find this variable. A quick fix is to
move it earlier, but probably it should be somewhere else entirely.