zsh: add some fzf widgets
They are heavily based on this:

https://github.com/junegunn/fzf/blob/master/shell/key-bindings.zsh

Since, however, it is highly unconfigurable (it just dictates mappings,
which I don't care for) and far too generic, and tricky to put in
version control without vendoring it, I've just reimplemented[1] it.
This will give me more flexibility to get funky with it later.

[1] here "reimplemented" means "I stole the code and removed everything
I could"