nvim: replace ultisnips with luasnips
UltiSnips had a long and productive life, but sadly, it's a mess to set
it up, and it's also a mess to get it to work in a bunch of situations.
I'm kicking it out for luasnip.

I still haven't figured out which snippets I'm going to use and how
exactly I'm going to use them, but hopefully, that will become evident
eventually.

I've put it back in my vim (as in not-NeoVim) config, but I haven't made
it works perfectly.