nvim: add ret snippet
ret expands to return 0, "", nil or whatever the default values of the
function that return is bound to.

I'm don't write enough go to be worth the time, but I wanted to learn
enough treesitter and luasnip to do something, and this was actually
cool. Inspired from the famous tj snippet, but a bit simpler.