nvim: fix lsp for lua a bit
I have broken snippet completion for function calls in lua. That is,
function calls expand to call_func instead of call_func(a, b), which is
nice.

I had to dig here to figure it out:

  https://github.com/sumneko/lua-language-server/wiki/Settings

I may have broken it in rust and go, but hopefully it's fine. I'm still
not sure how capabilities work.