nvim: all-buffer competion for ruby
Had to move to set hidden in order for that to work. I did it a bit
cheekily by just removing the override, which means that it's going to
be set hidden in neovim and set nohidden in vim.

Mange assures me it's the best thing ever, so I'm happy to have to try
it in nvim ;)