wezterm: use background color to indicate context
I'd like to know whether I'm in a particular context (as in, either [a]
a tmux session or [b] a non-default workspace in WezTerm). Since this is
where I usually am, I'd like that to be pure black. When I'm out of
either of those, I'd like the background to be colored differently.