nvim: get <D-e> to work with wezterm
I'm not supporting this in regular Vim, because it still uses nexus,
which now only benefits from the send-keys functionality (the running
bit is now in runner.vim).

I don't want to rewrite this in VimScript (so I chose lua), but I also
don't want to axe it from Vim8 yet (no good reason).